Replaced C-Style typedef struct {} foo; by C++-Style struct foo {}; everywhere
git-svn-id: https://svn.code.sf.net/p/extremetuxracer/code/trunk@313 0420edf4-82e4-42fc-9478-35b55e6d67a3

+++ env.cpp
@@ -141,7 +141,7 @@ bool CEnvironment::LoadEnvironmentList () {
 	}
 
 	locs.resize(list.Count());
-	for (int i=0; i<list.Count(); i++) {
+	for (size_t i=0; i<list.Count(); i++) {
 		string line = list.Line (i);
 		locs[i].name = SPStrN (line, "location", "");
 	}
